Restore missing CLI flags
These flags are missing in this new rebased version of the repository.
This code was previously merged. This is useful for testing the
tuple-grouping functionality without going all the way through the
process of linking with the Qwerty compiler. For example:

    $ bin/qir-xacc -i ../examples/qwerty/ghz5.ll -a qpp --group-tuples --no-print-accelbuf
    tuple ret length 5 distinct results 2
    tuple ret result 00000 count 474
    tuple ret result 11111 count 550
    tuple discarded length 0 distinct results 0

Here's the output of --help:

    $ bin/qir-xacc --help
    Usage: bin/qir-xacc [OPTIONS] input

    Positionals:
      input TEXT REQUIRED         QIR input file

    Options:
      -h,--help                   Print this help message and exit
      -i,--input TEXT REQUIRED    QIR input file
      -a,--accelerator TEXT REQUIRED
                                  Accelerator name
      -s,--shots INT [1024]       Number of shots
      --print-accelbuf,--no-print-accelbuf{false}
                                  Print XACC AcceleratorBuffer
      --group-tuples,--no-group-tuples{false}
                                  Print per-tuple measurement statistics rather than per-qubit

I wish it wouldn't print "{false}", but I read through the CLI11 source,
and I cannot find an easy way to turn that off.